+++ parsley-core/src/ghc/Parsley/Internal.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE PatternSynonyms, CPP #-}
+{-# LANGUAGE PatternSynonyms #-}
 {-|
 Module      : Parsley.Internal
 Description : The gateway into the internals: here be monsters!
@@ -16,35 +16,21 @@ module Parsley.Internal (
     module Primitives,
     module THUtils,
     module Trace,
-#if MIN_VERSION_parsley_core(2,0,0)
-#else
-    module Frontend,
-#endif
     module Backend,
     parse
   ) where
 
-#if MIN_VERSION_parsley_core(2,0,0)
 import Parsley.Internal.Backend  (codeGen, eval)
 import Parsley.Internal.Frontend (compile)
-#endif
 
 import Parsley.Internal.Backend         as Backend    (
     Input,
-#if MIN_VERSION_parsley_core(2,0,0)
-#else
-    codeGen, eval
-#endif
   )
 import Parsley.Internal.Core            as Core
 import Parsley.Internal.Core.Primitives as Primitives (
     pure, (<*>), (*>), (<*),
     (<|>), empty,
     satisfy, lookAhead, try, notFollowedBy,
-#if MIN_VERSION_parsley_core(2,0,0)
-#else
-    chainPre, chainPost,
-#endif
     loop,
     Reg, newRegister, get, put,
     conditional, branch,
@@ -52,10 +38,6 @@ import Parsley.Internal.Core.Primitives as Primitives (
     debug
   )
 import Parsley.Internal.Common.Utils    as THUtils    (Quapplicative(..), WQ, Code, makeQ)
-#if MIN_VERSION_parsley_core(2,0,0)
-#else
-import Parsley.Internal.Frontend        as Frontend   (compile)
-#endif
 import Parsley.Internal.Trace           as Trace      (Trace(trace))
 
 parse :: (Trace, Input input) => Parser a -> Code (input -> Maybe a)