+++ parsley-core/src/ghc/Parsley/Internal/Core.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE CPP #-}
 {-|
 Module      : Parsley.Internal.Core
 Description : The main AST and datatypes are found here
@@ -10,18 +9,10 @@ Stability   : unstable
 -}
 module Parsley.Internal.Core (
     Parser,
-#if MIN_VERSION_parsley_core(2,0,0)
-#else
-    ParserOps,
-#endif
     module Parsley.Internal.Core.Defunc,
     module Parsley.Internal.Core.InputTypes
   ) where
 
-import Parsley.Internal.Core.Defunc hiding (lamTerm, lamTermBool)
+import Parsley.Internal.Core.Defunc hiding (lamTerm)
 import Parsley.Internal.Core.InputTypes
-#if MIN_VERSION_parsley_core(2,0,0)
 import Parsley.Internal.Core.Primitives (Parser)
-#else
-import Parsley.Internal.Core.Primitives (Parser, ParserOps)
-#endif