+++ parsley-core/ChangeLog.md
@@ -131,3 +131,11 @@ Fix for issue #27
 * Added `PosSelector`
 * Added `line` and `col` to primitives and `Position` to Combinator AST, as well as `SelectPos` to instructions.
 * Changed `OFFSET` to `INPUT` in `Defunc`
+
+## 2.0.0.0 -- TBD
+
+* Removed `compile`, `eval`, `codeGen` from the API.
+* Removed `chainPre` and `chainPost` from the API.
+* Added `Typeable` constraint to `LIFTED`.
+* Removed `lamTermBool` and `userBool` from the API: `Typeable` subsumes them.
+* Added `RANGES` to the `Defunc` API.